Added support for plotting more than 9 topics
Before this patch, the subplots became smaller and smaller the more topics you added. Was unusable above 18 topics and threw errors when slightly larger than that.
Now the subplots are more or less the same size all the time. Also supports different amounts of columns, and changed default columns from 3 to 4 because it's a bit less wasteful. 3 columns still looks nice.