Make SklearnEmbedder to return a dense array
As is, the embed() method returns a `scipy.sparse._csr.csr_matrix` object which can't be used with `np.average()` directly that is used to get the topic embeddings. This is because `np.average()` try to cast the passed object using `np.asanyarray()` which returns it as an array with empty shape.

This fix simply uses the `.toarray()` method of `scipy.sparse._csr.csr_matrix` to return always a numpy array (dense).